+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 1.0.1
+
+- Adds [ScrollController] for showed & picked items.
+- Adds [ScrollPhysics] for showed & picked items.
+- Adds a choice whether to sort showed & picked items.
+- Adds picked item remove icon parameter.
+- Exposes picked item's [BoxDecoration].
+- Exposes clear all & select all buttons' animation curves.
+
 ## 1.0.0
 
 - Initial release.